requirements.txt
fastapi
PyMuPDF
pillow
pytesseract
requests
uvicorn[standard]
python-multipart